requirements.txt
mcp[cli]
httpx
pprintpp
openai
python-dotenv